POR 2688 - added quickbook timecard data to employee profile (#292)
* added quickbooks time data card

* refactored EmployeeInfoCard.vue and formatted emmployee page

* fixed styling on header and hireDate values bug

* 😫
